Update ReverseWord.java
Change to using BufferedReader and BufferedWriter.
Create print method and boolean variable to check wheter char is in tag or not.
if char is '<', boolean is true and following chars are printed in turns.
if char is '>', boolean is false.
if char is out of tag(normal word), it calls print method and print reverse or push to stack.